Input sensor alcohol & output LED also LCD
in this project im gonna show you how to detect an alcohol with proteus. in An alcohol sensor detects the attentiveness of alcohol gas in the air and an analog voltage is an output reading. The sensor can activate at temperatures ranging from -10 to 50° C with a power supply is less than 150 Ma to 5V. The sensing range is from 0.04 mg/L to 4 mg/L, which is suitable for breathalyzers.